Move dateslider component to common
