+++ app/scripts/components/common/dateslider/constants.ts
