+++ app/scripts/components/common/dateslider/context.tsx
