+++ app/scripts/components/common/dateslider/data-points.tsx
