+++ app/scripts/components/common/dateslider/date-axis.tsx
