+++ app/scripts/components/common/dateslider/faders.tsx
