+++ app/scripts/components/common/dateslider/index.tsx
