+++ app/scripts/components/common/dateslider/trigger-rect.tsx
