+++ app/scripts/components/common/dateslider/utils.tsx
